Add initial version of vmf2obj
vmf2obj can convert a Source engine .vmf file saved in Hammer to a .obj file which can be imported by Unreal Engine. It converts non-entity brushes to a mesh with no materials. Tools surfaces are skipped. I apologize in advance to anyone who reads the source code as it's very quick and dirty. Use a C++17 compiler. Tested on Windows with VS 2017 15.9.6.